Adding fac fxns and cost fxns
Including some print statements because I want to see the status of the gas as it goes through each stage. We can/will likely remove these as we get it workign and we can either store it in the facilities.py or remove it completely. I wrote it here because if i write it in the fac, then I don't know which stage we're at when the fxn prints it.